deploytool: split actions into their own files
